% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{.tcps}
\alias{.tcps}
\title{TCPS Shell}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 101 rows and 5 columns.}
\usage{
.tcps
}
\description{
TCPS Shell
}
\keyword{datasets}
