% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lever_scale}
\alias{lever_scale}
\title{Plot the likert scale for a TCPS lever}
\usage{
lever_scale(x, choice)
}
\arguments{
\item{x}{input tidy TCPS data}

\item{choice}{lever choice}
}
\value{
nify plot!
}
\description{
Plot the likert scale for a TCPS lever
}
