% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{likert_scale}
\alias{likert_scale}
\title{Create a likert scale plot for agreement and importance for a TCPS lever}
\usage{
likert_scale(x, choice)
}
\arguments{
\item{x}{input tidy TCPS data}

\item{choice}{lever choice}
}
\value{
nify plot!
}
\description{
Create a likert scale plot for agreement and importance for a TCPS lever
}
