% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{pal_one}
\alias{pal_one}
\title{TCPS ridgeline palette option #1 for use in tcps_lever_ridgeline}
\format{An object of class \code{character} of length 2.}
\usage{
pal_one
}
\description{
TCPS ridgeline palette option #1 for use in tcps_lever_ridgeline
}
\keyword{datasets}
