% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{pal_three}
\alias{pal_three}
\title{Ridgeline palette three}
\format{An object of class \code{character} of length 2.}
\usage{
pal_three
}
\description{
Ridgeline palette three
}
\keyword{datasets}
