% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale_helper}
\alias{scale_helper}
\title{Helper function to properly process and tidy TCPS data. Uses a tibble as input, and pmap to run the function}
\usage{
scale_helper(data, survey, lever, questions)
}
\arguments{
\item{data}{nested data column of the input tibble}

\item{survey}{survey type}

\item{lever}{lever choice}

\item{questions}{nested list of questions for each lever}
}
\value{
tidied tibble for each choice
}
\description{
Helper function to properly process and tidy TCPS data. Uses a tibble as input, and pmap to run the function
}
