% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale_likert}
\alias{scale_likert}
\title{Likert scale helper/contructor for lever_scale}
\usage{
scale_likert(x)
}
\arguments{
\item{x}{input data frame to process}
}
\value{
a tidy data frame of likert data for a scale
}
\description{
Likert scale helper/contructor for lever_scale
}
