% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{tcps_read_excel}
\alias{tcps_read_excel}
\title{Read the TCPS from a qualtrics export file or files, and transform the data into a tidy format}
\usage{
tcps_read_excel(file)
}
\arguments{
\item{file}{full file path(s) to an excel file(s) exported from the survey system}
}
\value{
a single data frame containing only TCPS data, excludes all additional questions added to the survey.
}
\description{
Read the TCPS from a qualtrics export file or files, and transform the data into a tidy format
}
