% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{tcps_report}
\alias{tcps_report}
\title{Generate TCPS report}
\usage{
tcps_report(path_to_data, name_of_school, format = "html")
}
\arguments{
\item{path_to_data}{full file path to the directory containing the TCPS data sets}

\item{name_of_school}{string of the name of the institution}

\item{format}{string denoting the output format, one of "html", "word" or "pdf"}
}
\value{
Report in provided format in current working directory
}
\description{
Generate TCPS report
}
