% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{tcps_sample}
\alias{tcps_sample}
\title{Retrieve file paths to TCPS sample data.}
\usage{
tcps_sample(path = NULL, ...)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}

\item{...}{Additional parameters pass on to dir}
}
\description{
tcps comes bundled with some example files in its \code{inst/extdata}
directory. This function make them easy to access by passing the sample name
as a quoted string.  It can be one of "staff_sample.xlsx", "student_sample.xlsx" or
"faculty_sample.xlsx"
}
