% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{tcps_tidy_spss}
\alias{tcps_tidy_spss}
\title{Read the TCPS data contained within an SPSS file generated the TCPS group and transform the data into a tidy format}
\usage{
tcps_tidy_spss(file)
}
\arguments{
\item{file}{full file path to the SPSS .sav file}
}
\value{
a tidy data frame of TCPS data
}
\description{
Read the TCPS data contained within an SPSS file generated the TCPS group and transform the data into a tidy format
}
