% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_tcps}
\alias{theme_tcps}
\title{A Custom ggplot2 theme for the TCPS}
\usage{
theme_tcps(
  base_family = "mono",
  base_size = 11,
  strip_text_family = base_family,
  strip_text_size = 12,
  plot_title_family = "mono",
  plot_title_size = 16,
  plot_title_margin = 10,
  subtitle_family = "mono",
  subtitle_size = 12,
  subtitle_margin = 15,
  caption_family = "serif",
  caption_size = 9,
  caption_margin = 10,
  axis_title_family = "serif",
  axis_title_size = 9,
  axis_title_just = "mm",
  grid = TRUE,
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family}{base font family}

\item{base_size}{base font size}

\item{strip_text_family}{facet label font family}

\item{strip_text_size}{facet label text size}

\item{plot_title_family}{plot tilte family}

\item{plot_title_size}{plot title font size}

\item{plot_title_margin}{plot title margin}

\item{subtitle_family}{plot subtitle family}

\item{subtitle_size}{plot subtitle size}

\item{subtitle_margin}{plot subtitle margin}

\item{caption_family}{plot caption family}

\item{caption_size}{plot caption size}

\item{caption_margin}{plot caption margin}

\item{axis_title_family}{axis title font family}

\item{axis_title_size}{axis title font size}

\item{axis_title_just}{axis title font justification \code{blmcrt}}

\item{grid}{panel grid (\code{TRUE}, \code{FALSE}, or a combination of
\code{X}, \code{x}, \code{Y}, \code{y})}

\item{axis}{axis \code{TRUE}, \code{FALSE}, \code{xy}}

\item{ticks}{ticks}
}
\description{
Custom ggplot2 theme for the TCPS Package
heavy credits for influencing the theme function
go to @hrbrmstr (Bob Rudis)
}
\details{
\url{https://www.google.com/fonts}
}
