% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{tidy_tcps_windsor}
\alias{tidy_tcps_windsor}
\title{Read the TCPS data contained within an SPSS file generated by Ken Meadows and transform the data into a tidy format}
\usage{
tidy_tcps_windsor(file)
}
\arguments{
\item{file}{full file path to the SPSS .sav file}
}
\value{
a tidy data frame of TCPS data
}
\description{
Read the TCPS data contained within an SPSS file generated by Ken Meadows and transform the data into a tidy format
}
