% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lever_ridgeline}
\alias{lever_ridgeline}
\title{Produce ridgeline plots for a TCPS lever.}
\usage{
lever_ridgeline(x, name = "University Name", lever = NULL,
  pal = pal_one, aggregate = FALSE)
}
\arguments{
\item{x}{input tidy TCPS data frame}

\item{name}{name of institution}

\item{lever}{short name of the lever of interest}

\item{pal}{a simple two color palette for agreement/importance in the plot.}

\item{aggregate}{TRUE/FALSE.  If TRUE, aggregate responses across all populations (surveys) in the data frame}
}
\value{
a nifty plot!
}
\description{
Produce ridgeline plots for a TCPS lever.
}
