% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lever_scale}
\alias{lever_scale}
\title{Produce the lever scale item visualization}
\usage{
lever_scale(x, choice)
}
\arguments{
\item{x}{the input tidy tcps data frame}

\item{choice}{the short name for the lever scale}
}
\value{
nify plot
}
\description{
Produce the lever scale item visualization
}
