% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{pal_three}
\alias{pal_three}
\title{TCPS ridgeline palette option #2 for use in tcps_lever_ridgeline}
\format{An object of class \code{character} of length 2.}
\usage{
pal_three
}
\description{
TCPS ridgeline palette option #2 for use in tcps_lever_ridgeline
}
\keyword{datasets}
