% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tcps_lever_scale}
\alias{tcps_lever_scale}
\title{Visualize the likert scale items comprising a lever}
\usage{
tcps_lever_scale(x, choice, name = "University Name")
}
\arguments{
\item{x}{the input tidy tcps data frame}

\item{choice}{the short name for the lever scale}

\item{name}{name of the institution}
}
\value{
nify plot
}
\description{
Visualize the likert scale items comprising a lever
}
