% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcps_sample}
\alias{tcps_sample}
\title{Sample simulated data for the TCPS survey}
\format{A data frame with 448 rows and 46 variables:}
\usage{
tcps_sample
}
\description{
A dataset containing simulated raw data resulting from tidy_tcps
}
\keyword{datasets}
