% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{tcps_summary_stats}
\alias{tcps_summary_stats}
\title{Calculate basic summary statistics for a tidied data frame of TCPS data}
\usage{
tcps_summary_stats(data)
}
\arguments{
\item{data}{a tidy tcps data frame from tcps_read_excel()}
}
\value{
a data frame of summary statistics for each scale, survey group and question
}
\description{
Calculate basic summary statistics for a tidied data frame of TCPS data
}
